requirements.txt
tqdm
wikipedia-api